refactor: account for hf models misalignment
