+++ spacy_llm/backends/__init__.py
@@ -1,4 +1,4 @@
-from .hf import backend_hf
+from .hf import backend_dolly_hf
 from .langchain import backend_langchain, query_langchain
 from .minichain import backend_minichain, query_minichain
 from .rest import backend_rest
