+++ spacy_llm/backends/hf/__init__.py
@@ -1,0 +1,1 @@
+from .dolly import backend_dolly_hf
